pyproject.toml
[tool.poetry]
name = "smarch"
version = "0.1.0"
description = "AWS Lambda based SMB file archiver to S3"
authors = ["Masanori-Yokoyama"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.86"
pysmb = "^1.2.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=smarch --cov-report=term-missing"

[tool.coverage.run]
source = ["smarch"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
